+++ include/snn/trainer/training_process/GD_adagrad_process.h
@@ -20,7 +20,7 @@ namespace SNN
 	class GD_adagrad_process : public Trainig_process
 	{
 	public:
-		GD_adagrad_process(Perceptron* p_perceptron, double p_step);
+		GD_adagrad_process(Perceptron* p_perceptron, float p_step);
 		~GD_adagrad_process();
 
 		void propagate(vector<Trainig_process*>* process, bool out = false);
@@ -28,8 +28,8 @@ namespace SNN
 		static void set_default_configuration(trainig_config_t* configuration);
 
 	private:
-		double m_step;
-		double ss_gradient;
+		float m_step;
+		float ss_gradient;
 	};
 } // SNN_network
