+++ include/snn/trainer/training_process/GD_nesterov_process.h
@@ -20,7 +20,7 @@ namespace SNN
 	class GD_nesterov_process : public Trainig_process
 	{
 	public:
-		GD_nesterov_process(Perceptron* p_perceptron, double p_step, double momentum_factor);
+		GD_nesterov_process(Perceptron* p_perceptron, float p_step, float momentum_factor);
 		~GD_nesterov_process();
 
 		void propagate(vector<Trainig_process*>* process, bool out = false);
@@ -28,11 +28,11 @@ namespace SNN
 		static void set_default_configuration(trainig_config_t* configuration);
 
 	private:
-		double m_step;
+		float m_step;
 
-		double m_momentum_factor;
+		float m_momentum_factor;
 
-		double m_delta_1;
+		float m_delta_1;
 	};
 } // SNN_network
