+++ include/snn/trainer/training_process/GD_rmsprop_process.h
@@ -20,7 +20,7 @@ namespace SNN
 	class GD_rmsprop_process : public Trainig_process
 	{
 	public:
-		GD_rmsprop_process(Perceptron* p_perceptron, double p_step);
+		GD_rmsprop_process(Perceptron* p_perceptron, float p_step);
 		~GD_rmsprop_process();
 
 		void propagate(vector<Trainig_process*>* process, bool out = false);
@@ -28,9 +28,9 @@ namespace SNN
 		static void set_default_configuration(trainig_config_t* configuration);
 
 	private:
-		double m_step;
+		float m_step;
 
-		double m_Eg;
+		float m_Eg;
 	};
 } // SNN_network
