+++ include/snn/trainer/training_process/Steepest_descent_process.h
@@ -20,7 +20,7 @@ namespace SNN
 	class Steepest_descent_process : public Trainig_process
 	{
 	public:
-		Steepest_descent_process(Perceptron* p_perceptron, double p_step);
+		Steepest_descent_process(Perceptron* p_perceptron, float p_step);
 		~Steepest_descent_process();
 
 		void propagate(vector<Trainig_process*>* process, bool out = false);
@@ -28,7 +28,7 @@ namespace SNN
 		static void set_default_configuration(trainig_config_t* configuration);
 
 	private:
-		double m_step;
+		float m_step;
 	};
 
 } // namespace SNN_network
