+++ include/snn/trainer/training_process/Trainig_process.h
@@ -26,27 +26,27 @@ namespace SNN
 		Trainig_process(Perceptron* p_perceptron);
 		virtual ~Trainig_process() {};
 
-		virtual void set_error(double T);
+		virtual void set_error(float T);
 		virtual void propagate(vector<Trainig_process*>* process, bool out = false) { process; out; };
 		virtual void compute();
 
-		virtual void add_to_gradient(double value) { m_gradient += value; };
+		virtual void add_to_gradient(float value) { m_gradient += value; };
 
 	protected:
 		Perceptron* m_perceptron;
 
-		double m_bia_gradient;
-		vector<double> m_w_gradient;
+		float m_bia_gradient;
+		vector<float> m_w_gradient;
 
-		double m_gradient;
-		double m_error;
-		double m_delta;
+		float m_gradient;
+		float m_error;
+		float m_delta;
 
 		void derivate_perceptron() { m_perceptron->derivate(); };
-		double get_derivate() { return *(m_perceptron->m_derivate.begin()); };
-		double get_single_derivate() { return m_perceptron->derivate_single(); };
+		float get_derivate() { return *(m_perceptron->m_derivate.begin()); };
+		float get_single_derivate() { return m_perceptron->derivate_single(); };
 
-		void add_to_precedent(vector<Trainig_process*>* process, double factor);
+		void add_to_precedent(vector<Trainig_process*>* process, float factor);
 	};
 
 } // namespace SNN_network
